Rename "DLLP" to "Frame"
While the TLP sequence number and LCRC are added at the Data Link Layer,
a "Data Link Layer Packet" is less of a container for TLPs and more of a
packet that "targets" the Data Link Layer. Furthermore, the start and
end tags are really more part of the Physical Layer, despite the Data
Link Layer being the layer that specifies which symbols should be added
into the stream.

So it's more like TLPs and DLLPs are sent over the implicit "framing"
layer, which is implemented by what the spec calls the "Logical
Sub-block" of the Physical Layer.