Add note for future enhancement
Since this data is coming from a protocol analyzer, it should not be
assumed that any of the packet data is valid. And since the data is
received sequentially, any error in the stream will likely result in the
remainder of the packet becoming corrupted, so decoding backwards from
the end of the valid data is not likely to decode properly. Therefore,
it's best to decode all fields sequentially, and we should make sure the
dissector is able to recover gracefully when it encounters invalid data.