Remove strict ordering requirement for TLP completion tracking
Since the transaction ID is only used as a key until a matching
completion is found, we don't need to make every transaction ID globally
unique. Removing the code that does this has the side-benefit of
enabling tracking of requests/completions where the requester uses tags
opportunistically, as soon as they're freed by a completion, instead of
sequentially. And removing the map insertion and two map lookups offers
a slight performance boost as well.