Factor out PCIe capture dissector from standard PCIe protocol dissectors
Since the code was starting to get a bit difficult to manage, I decided
to factor out the non-standard / Agilent-specific parts of the code.
This puts it on the same level as the NetTLP dissector, which also lives
in a separate file and calls a standard PCIe dissector where necessary.